`storage`: add `map_building_reducer`
This reducer builds a `key_offset_map` for a segment starting from an offset
`start_offset_inclusive`, until the memory limit of the map is reached
or the segment is fully indexed. `end_of_stream()` returns a boolean value
indicating whether the segment was fully indexed or not, and the map can be
probed for its `max_offset()` in order to start the next round of indexing
(using `start_offset_inclusive = model::next_offset(max_offset())`).

It is intended to be used in the case that a segment was not able to be fully
indexed during a round of sliding window compaction, falling back instead
to a "chunked" compaction routine.

This reducer will be used during the "chunked" compaction procedure by
starting from the base offset in the segment, and performing as many rounds
of map building and de-duplication as is required until `end_of_stream()` returns
`true`.