dl/coordinator: add snapshot remover
Introduces a new abstraction to remove old snapshots from a given table.
This effectively reduces the window with which users can perform
point-in-time queries. It is intended that this operation is standalone
and doesn't require additional changes to the coordinator STM to run.

This commit only introduces the remover and an no-op implementation for
ease of testing, similar to the file_committer abstraction. A subsequent
change will add this to the coordinator.