r/types: do not use reader in append entries request
`model::record_batch_reader` was not the best abstraction to represent
the batches in `append_entries_request`. Number of batches contained in
the request is known upfront and they are always materialized in memory.
Therefore it is more convenient and easier to store batches as a plain
`chunked_vector` instead of the reader.

Signed-off-by: Michał Maślanka <michal@redpanda.com>