storage: allow buffering of multiple spill keys
This is effectively a refactor commit which splits the spill(key)
interface into separate functions:

Prior to this commit the spill key index had the spill(key) interface
which would accept a single key. This commit introduces a spill_payload
type which can be used to collect many spilled keys and then spill them
all at once with a spill(payload) interface.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>