storage: try to avoid spill key index erase hot spot
Try to start spilling from a dynamic location in the index based on the
last inserted key as a seed. The actual removal starts after the last
inserted key which will map to a random key and (hopefully) not a key
that has temporal locality with recently inserted items.

If a index receives a high number of duplicates then the removal should
be no worse than previous repeated removal from begin().

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>