crash_tracker: custom `crash_loop_limit_reached`
Define a custom exception for `crash_loop_limit_reached` to allow
pattern matching for it in a follow up commit.

We specifically do not want to record `crash_loop_limit_reached` errors
as crashes because they are generally not informative crashes. Recording
them as real crashes would build up garbage on disk, which would lead to
real crash logs expiring on disk earlier.