crash_tracker: introduce `prepared_writer`
Introduces the core stateful writer which is used to write out
`crash_description` serde objects to crash report files. It is able to
write out the crash reports in an async-signal-safe way by:
 * using only async signal safe syscalls
 * pre-allocating all the memory necessary to construct, serialize and
   write our `crash_description` objects to disk
 * not allocating in fill() and write(), i.e. on the signal handler path

Ref: https://man7.org/linux/man-pages/man7/signal-safety.7.html

The `prepared_writer` is hooked up to the `recorder` in a later commit
of this PR.

Note that fill() may race with other calls of fill() or with release(),
and the class safely breaks this race. These races can occur if multiple
signals arrive on different threads at the same time, or if a signal
arrives concurrently with release() being called on shard 0.