tests/partition_reassignments_test/verification: process 3.9.0 output
This test processes output of kafka cli and fails if it differs from
expected. Kafka cli partition reassignment tool has been rewritten from
Scala into Java between versions 3.0.0 and 3.6.0, and now we use 3.9.0.
It's not that its output has changed too much, seems like an extra
newline was added or something like this. Rewrite the original output
processing in the test, so that it addresses the issue and is generally
more robust. Re-enable the test too.