dl/catalog_schema_manager: Wire up compatibility checks
Introduce check_schema_compat, called on the ensure_schema path, to call into
the compatibility layer. Returns an error, or an indication that the schema
changed, or nothing (no schema change).

Also update fill_field_ids to use the compat machinery to support filling field
IDs in the presence of type changes, changes in field order, or records with
missing fields. Errors iff not all destination fields could be filled.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>