r/buffered_protocol: disable buffering in raft buffered protocol
Changed the default value of a buffer size in raft buffered protocol to
disable the buffering completely. With the buffer enabled backpressure
from the followers wasn't propagated leading to a large queue of
requests waiting in the buffer. This made the replicate batcher to batch
far less effectively leading to increased latency and overall worse
performance.

We may reconsider enabling the buffer back if we implement different
batching strategy (a timer based) than batching only if backpressure is
propagated.

Signed-off-by: Michał Maślanka <michal@redpanda.com>