raft: dispatch heartbeats in dedicated scheduling group
Heartbeat requests servers as a timeout based failure detector for Raft
layer. Delivering and handling heartbeat requests is critical for the
cluster stability. Failing heartbeat delivery cause leader election and
initiate other resource consuming events. By running heartbeats in a
separate scheduling group with high priority we are making sure the
heartbeats are processed even if the cluster is overloaded.

Signed-off-by: Michał Maślanka <michal@redpanda.com>